+++ shelter/__init__.py
@@ -3,4 +3,4 @@ Simple Python's Tornado wrapper which provides helpers for creating
 a new project, writing management commands, service processes, ...
 """
 
-__version__ = '1.2.0.dev2'
+__version__ = '2.0.0.dev1'